+++ tidypolars/tidypolars.py
@@ -1,35 +1,25 @@
 import polars as pl
-from polars import col, Series, when
+from polars import col, Expr, Series, when
 import functools as ft
+from typing import Dict, List, Union
 
-from typing import Union, List
+from .funs import *
 
-def as_tibble(df):
-    df.__class__ = tibble
+def _as_Tibble(df):
+    df.__class__ = Tibble
     return df
 
-def col_expr(x):
-    if isinstance(x, pl.Expr):
-        return x
-    elif isinstance(x, str):
-        return col(x)
-    else:
-       raise ValueError("Invalid input for column selection") 
-
-#  Wrap all str inputs in col()  
-def col_exprs(x):
-    if is_list_like(x):
-        return [col_expr(val) for val in x]
-    else:
-        return [col_expr(x)]
+def _as_DataFrame(df):
+    df.__class__ = pl.DataFrame
+    return df
   
-def is_list_like(x):
+def _is_list_like(x):
     if isinstance(x, list) | isinstance(x, pl.Series):
         return True
     else:
         return False
 
-def as_list(x):
+def _as_list(x):
     if isinstance(x, list):
         return x.copy()
     elif isinstance(x, str):
@@ -37,85 +27,89 @@ def as_list(x):
     else:
         return list(x)
 
-def args_as_list(x):
+def _args_as_list(x):
     if len(x) == 0:
         return []
     elif isinstance(x[0], list):
         return x[0]
-    elif is_list_like(x[0]):
+    elif isinstance(x[0], pl.Series):
         return list(x[0])
     else:
         return [*x]
 
 # Convert kwargs to col() expressions with alias
-def kwargs_as_exprs(kwargs):
+def _kwargs_as_exprs(kwargs):
     return [expr.alias(key) for key, expr in kwargs.items()]
 
-class tibble(pl.DataFrame):
-    def arrange(self, *args, desc: Union[bool, List[bool]] = False) -> "tp.tibble":
+def _no_groupby(gb):
+    if isinstance(gb, Expr) | isinstance(gb, str) | isinstance(gb, list):
+        return False
+    else:
+        return True
+
+class Tibble(pl.DataFrame):
+    def __repr__(self) -> str:
+        df = _as_DataFrame(self)
+        return df.__str__()
+
+    def arrange(self, *args, desc: bool = False):
         """
         Arrange/sort rows
 
         Parameters
         ----------
-        *args : Union[str, Expr]
+        *args : 
             Columns to sort by
-
-        desc : Union[bool, List[bool]] = False
+        desc : 
             Should columns be ordered in descending order
 
         Examples
         --------
-        df = tp.tibble({'x': ['a', 'a', 'b'], 'y': range(3)})
-        
-        # Arrange in ascending order
-        df.arrange('x', 'y')
-        
-        # Arrange some columns descending
-        df.arrange('x', 'y', desc = [True, False])
+        >>> df = tp.Tibble({'x': ['a', 'a', 'b'], 'y': range(3)})
+        >>> # Arrange in ascending order
+        >>> df.arrange('x', 'y')
+        ...
+        >>> # Arrange some columns descending
+        >>> df.arrange('x', 'y', desc = [True, False])
         """
-        exprs = args_as_list(args)
-        return self.sort(exprs, reverse = desc).pipe(as_tibble)
+        exprs = _args_as_list(args)
+        return self.sort(exprs, reverse = desc).pipe(_as_Tibble)
 
-    def bind_cols(self, df: "tp.tibble") -> "tp.tibble" :
+    def bind_cols(self, df: "tp.Tibble"):
         """
         Bind data frames by columns
 
         Parameters
         ----------
-        df : tibble
+        df : Tibble
             Data frame to bind
 
         Examples
         --------
-        df1 = tp.tibble({'x': ['a', 'a', 'b'], 'y': range(3)})
-        df2 = tp.tibble({'a': ['c', 'c', 'c'], 'b': range(4, 7)})
-
-        df1.bind_cols(df2)
+        >>> df1 = tp.Tibble({'x': ['a', 'a', 'b'], 'y': range(3)})
+        >>> df2 = tp.Tibble({'a': ['c', 'c', 'c'], 'b': range(4, 7)})
+        >>> df1.bind_cols(df2)
         """
-        # TODO: Allow to work on multiple inputs
-        return self.hstack(df).pipe(as_tibble)
+        return self.hstack(df).pipe(_as_Tibble)
     
-    def bind_rows(self, df: "tp.tibble") -> "tp.tibble":
+    def bind_rows(self, df: "tp.Tibble"):
         """
         Bind data frames by row
 
         Parameters
         ----------
-        df : tibble
+        df : Tibble
             Data frame to bind
 
         Examples
         --------
-        df1 = tp.tibble({'x': ['a', 'a', 'b'], 'y': range(3)})
-        df2 = tp.tibble({'x': ['c', 'c', 'c'], 'y': range(4, 7)})
-
-        df1.bind_rows(df2)
+        >>> df1 = tp.Tibble({'x': ['a', 'a', 'b'], 'y': range(3)})
+        >>> df2 = tp.Tibble({'x': ['c', 'c', 'c'], 'y': range(4, 7)})
+        >>> df1.bind_rows(df2)
         """
-        # TODO: Allow to work on multiple inputs
-        return self.vstack(df).pipe(as_tibble)
+        return self.vstack(df).pipe(_as_Tibble)
 
-    def distinct(self, *args) -> "tp.tibble":
+    def distinct(self, *args):
         """
         Select distinct/unique rows
 
@@ -133,16 +127,21 @@ class tibble(pl.DataFrame):
         df.distinct('b')
         """
         # TODO: Create for series
-        args = args_as_list(args)
+        args = _args_as_list(args)
 
         if len(args) == 0:
             df = self.drop_duplicates()
         else:
             df = self.select(args).drop_duplicates()
         
-        return df.pipe(as_tibble)
+        return df.pipe(_as_Tibble)
+    
+    def head(self, n = 5, *args, groupby = None):
+        """Alias for .slice_head()"""
+        return self.slice_tail(n, groupby = groupby)
 
-    def filter(self, *args) -> "tp.tibble":
+    def filter(self, *args,
+               groupby: Union[str, Expr, List[str], List[Expr]] = None):
         """
         Filter rows on one or more conditions
 
@@ -151,63 +150,62 @@ class tibble(pl.DataFrame):
         *args : Expr
             Conditions to filter by
 
+        groupby : Union[str, Expr, List[str], List[Expr]]
+            Columns to group by
+
         Examples
         --------
-        df = tp.tibble({'a': range(3), 'b': ['a', 'a', 'b']})
-        
-        df.filter(col('a') < 2, col('c') == 'a')
-
-        df.filter((col('a') < 2) & (col('c') == 'a'))
+        >>> df = tp.Tibble({'a': range(3), 'b': ['a', 'a', 'b']})
+        >>> df.filter(col('a') < 2, col('c') == 'a')
+        >>> df.filter((col('a') < 2) & (col('c') == 'a'))
+        >>> df.filter(col('a') <= col('a').mean(),
+        ...           groupby = 'b')
         """
-        args = args_as_list(args)
+        df = _as_DataFrame(self)
+        args = _args_as_list(args)
         exprs = ft.reduce(lambda a, b: a & b, args)
-        return super().filter(exprs).pipe(as_tibble)
-    
-    def group_by(self, *args):
-        """
-        Group by one or more variables
 
-        Parameters
-        ----------
-        *args : Expr
-            Conditions to filter by
-
-        Examples
-        --------
-        df = tp.tibble({'a': range(3), 'b': range(3), 'c': ['a', 'a', 'b']})
+        if _no_groupby(groupby):
+            df = df.filter(exprs)
+        else:
+            df = df.groupby(groupby).apply(lambda x: x.filter(exprs))
         
-        df.group_by('c')
+        return df.pipe(_as_Tibble)
 
-        df.group_by('a', 'c')
-        """
-        args = args_as_list(args)
-        df = self.groupby(args)
-        df.__class__ = grouped_tibble
-        return df
-    
-    def mutate(self, *args, **kwargs) -> "tp.tibble":
+    def mutate(self, *args,
+               groupby: Union[str, Expr, List[str], List[Expr]] = None,
+               **kwargs):
         """
         Add or modify columns
 
         Parameters
         ----------
+        *args : Expr
+            Column expressions to add or modify
+
+        groupby : Union[str, Expr, List[str], List[Expr]]
+            Columns to group by
+        
         **kwargs : Expr
             Column expressions to add or modify
 
         Examples
         --------
-        df = tp.tibble({'a': range(3), 'b': range(3)})
-
-        df.mutate(double_a = col('a') * 2,
-                  a_plus_b = col('a') + col('b'))
-        
-        df.mutate((col(['a', 'b]) * 2).prefix('double_'),
-                  a_plus_b = col('a') + col('b'))
+        >>> df = tp.Tibble({'a': range(3), 'b': range(3)})
+        >>> df.mutate(double_a = col('a') * 2,
+        ...           a_plus_b = col('a') + col('b'))
+        >>> df.mutate((col(['a', 'b]) * 2).prefix('double_'),
+        ...           a_plus_b = col('a') + col('b'))
         """
-        exprs = args_as_list(args) + kwargs_as_exprs(kwargs)
-        return self.with_columns(exprs).pipe(as_tibble)
+        exprs = _args_as_list(args) + _kwargs_as_exprs(kwargs)
+        if _no_groupby(groupby):
+            out = self.with_columns(exprs)
+        else:
+            out = self.groupby(groupby).apply(lambda x: x.with_columns(exprs))
+        
+        return out.pipe(_as_Tibble)
 
-    def pull(self, var = None):
+    def pull(self, var: str = None):
         """
         Extract a column as a series
 
@@ -218,16 +216,15 @@ class tibble(pl.DataFrame):
 
         Examples
         --------
-        df = tp.tibble({'a': range(3), 'b': range(3))
-        
-        df.pull('a')
+        >>> df = tp.Tibble({'a': range(3), 'b': range(3))
+        >>> df.pull('a')
         """
         if var == None:
             var = self.columns[-1]
         
         return self.get_column(var)
     
-    def relocate(self, *args, before: str = None, after: str = None) -> "tp.tibble":
+    def relocate(self, *args, before: str = None, after: str = None):
         """
         Move a column or columns to a new position
 
@@ -238,11 +235,9 @@ class tibble(pl.DataFrame):
 
         Examples
         --------
-        df = tp.tibble({'a': range(3), 'b': range(3), 'c': ['a', 'a', 'b']})
-        
-        df.relocate('a', before = 'c')
-
-        df.relocate('b', after = 'c')
+        >>> df = tp.Tibble({'a': range(3), 'b': range(3), 'c': ['a', 'a', 'b']})
+        >>> df.relocate('a', before = 'c')
+        >>> df.relocate('b', after = 'c')
         """
         move_cols = pl.Series(list(args))
 
@@ -281,7 +276,24 @@ class tibble(pl.DataFrame):
 
             return self.select(ordered_cols)
     
-    def select(self, *args) -> "tp.tibble":
+    def rename(self, mapping: Dict[str, str]):
+        """
+        Rename columns
+
+        Parameters
+        ----------
+        *args : Dict[str, str]
+            Dictionary mapping of new names
+
+        Examples
+        --------
+        >>> df = tp.Tibble({'x': range(3), 't': range(3), 'z': ['a', 'a', 'b']})
+        >>> df.rename({'x': 'new_x'})
+        """
+        df = _as_DataFrame(self)
+        return df.rename(mapping).pipe(_as_Tibble)
+    
+    def select(self, *args):
         """
         Select or drop columns
 
@@ -292,92 +304,133 @@ class tibble(pl.DataFrame):
 
         Examples
         --------
-        df = tp.tibble({'a': range(3), 'b': range(3), 'c': ['a', 'a', 'b']})
-        
-        df.select('a', 'b')
+        >>> df = tp.Tibble({'a': range(3), 'b': range(3), 'c': ['a', 'a', 'b']})
+        >>> df.select('a', 'b')
+        >>> df.select(col('a'), col('b'))
+        """
+        args = _args_as_list(args)
+        return super().select(args).pipe(_as_Tibble)
 
-        df.select(col('a'), col('b'))
+    def slice(self, *args, groupby = None):
         """
-        args = args_as_list(args)
-        return super().select(args).pipe(as_tibble)
-    
-    def summarize(self, *args, **kwargs) -> "tp.tibble":
-        """
-        Aggregate data with summary statistics
+        Grab rows from a data frame
 
         Parameters
         ----------
-        **kwargs : Expr
-            Column expressions to add or modify
+        *args : Union[int, List[int]]
+            Rows to grab
+        
+        groupby : Union[str, Expr, List[str], List[Expr]]
+            Columns to group by
 
         Examples
         --------
-        df = tp.tibble({'a': range(3), 'b': range(3), 'c': ['a', 'a', 'b']})
+        >>> df = tp.Tibble({'a': range(3), 'b': range(3), 'c': ['a', 'a', 'b']})
+        >>> df.slice(0, 1)
+        >>> df.slice(0, groupby = 'c')
+        """
+        rows = _args_as_list(args)
+        if _no_groupby(groupby):
+            df = self[rows]
+        else:
+            df = self.groupby(groupby).apply(lambda x: x[rows])
+        return df.pipe(_as_Tibble)
+
+    def slice_head(self, n: int = 5, *args, groupby = None):
+        """
+        Grab top rows from a data frame
+
+        Parameters
+        ----------
+        n : int
+            Number of rows to grab
         
-        df.summarize(avg_a = col('a').mean())
+        *args :
+            Currently unused
+        
+        groupby : Union[str, Expr, List[str], List[Expr]]
+            Columns to group by
 
-        df.summarize(avg_a = col('a').mean(),
-                     max_b = col('b').max()))
+        Examples
+        --------
+        >>> df = tp.Tibble({'a': range(3), 'b': range(3), 'c': ['a', 'a', 'b']})
+        >>> df.slice_head(2)
+        >>> df.slice_head(1, groupby = 'c')
         """
-        exprs = args_as_list(args) + kwargs_as_exprs(kwargs)
-        return super().select(exprs).pipe(as_tibble)
+        df = _as_DataFrame(self)
+        args = _args_as_list(args)
+        col_order = df.columns
+        if _no_groupby(groupby):
+            df = df.head(n)
+        else:
+            df = df.groupby(groupby).head(n)
+        return df.pipe(_as_Tibble).select(col_order)
 
-class grouped_tibble(pl.eager.frame.GroupBy):
-    def filter(self, *args) -> "tp.tibble":
+    def slice_tail(self, n: int = 5, *args, groupby = None):
         """
-        Filter rows on one or more conditions by group
+        Grab bottom rows from a data frame
+
+        Parameters
+        ----------
+        n : int
+            Number of rows to grab
+        
+        *args :
+            Currently unused
+        
+        groupby : Union[str, Expr, List[str], List[Expr]]
+            Columns to group by
+
+        Examples
+        --------
+        >>> df = tp.Tibble({'a': range(3), 'b': range(3), 'c': ['a', 'a', 'b']})
+        >>> df.slice_tail(2)
+        >>> df.slice_tail(1, groupby = 'c')
+        """
+        df = _as_DataFrame(self)
+        args = _args_as_list(args)
+        col_order = df.columns
+        if _no_groupby(groupby):
+            df = df.tail(n)
+        else:
+            df = df.groupby(groupby).tail(n)
+        return df.pipe(_as_Tibble).select(col_order)
+    
+    def summarize(self, *args,
+                  groupby: Union[str, Expr, List[str], List[Expr]] = None,
+                  **kwargs):
+        """
+        Aggregate data with summary statistics
 
         Parameters
         ----------
         *args : Expr
-            Conditions to filter by
 
-        Examples
-        --------
-        df = tp.tibble({'a': range(3), 'b': ['a', 'a', 'b']})
-        
-        df.group_by('b').filter(col('a') < col('a').mean())
-        """
-        args = args_as_list(args)
-        exprs = ft.reduce(lambda a, b: a & b, args)
-        return self.apply(lambda df: df.filter(exprs)).pipe(as_tibble)
-    
-    def mutate(self, *args, **kwargs) -> "tp.tibble":
-        """
-        Add or modify columns by group
+        groupby : Union[str, Expr, List[str], List[Expr]]
+            Columns to group by
 
-        Parameters
-        ----------
         **kwargs : Expr
             Column expressions to add or modify
 
         Examples
         --------
-        df = tp.tibble({'a': range(3), 'b': ['a', 'a', 'b']})
+        >>> df = tp.Tibble({'a': range(3), 'b': range(3), 'c': ['a', 'a', 'b']})
+        >>> df.summarize(avg_a = col('a').mean())
+        >>> df.summarize(avg_a = col('a').mean(),
+        ...              groupby = 'c')
+        >>> df.summarize(avg_a = col('a').mean(),
+        ...              max_b = col('b').max())
+        """
+        exprs = _args_as_list(args) + _kwargs_as_exprs(kwargs)
+        df = _as_DataFrame(self)
+
+        if _no_groupby(groupby):
+            out = df.select(exprs)
+        else:
+            out = df.groupby(groupby).agg(exprs)
         
-        df.group_by('b').mutate(avg_a = col('a').mean())
-        """
-        exprs = args_as_list(args) + kwargs_as_exprs(kwargs)
-        return self.apply(lambda df: df.with_columns(exprs)).pipe(as_tibble)
+        return out.pipe(_as_Tibble)
 
-    def summarize(self, *args, **kwargs) -> "tp.tibble":
-        """
-        Aggregate data with summary statistics
-
-        Parameters
-        ----------
-        **kwargs : Expr
-            Column expressions to add or modify
-
-        Examples
-        --------
-        df = tp.tibble({'a': range(3), 'b': range(3), 'c': ['a', 'a', 'b']})
-        
-        df.summarize(avg_a = col('a').mean())
-
-        df.group_by('c').summarize(avg_a = col('a').mean(),
-                                   max_b = col('b').max())
-        """
-        exprs = args_as_list(args) + kwargs_as_exprs(kwargs)
-        return self.agg(exprs).pipe(as_tibble)
-
+    def tail(self, n = 5, *args, groupby = None):
+        """Alias for `.slice_tail()`"""
+        return self.slice_tail(n, groupby = groupby)